// Cognexia-Frontend-CRM_Frontend/backend/modules/05-inventory/src/entities/InventoryItem.d.ts
import { ItemCategory, ItemStatus, UnitOfMeasure } from '../enums';
import { StockTransaction } from './StockTransaction';
import { StockLocation } from './StockLocation';
import { CycleCount } from './CycleCount';
import { InventoryAdjustment } from './InventoryAdjustment';
import { InventoryAlert } from './InventoryAlert';
import { ReorderPoint } from './ReorderPoint';
export declare class InventoryItem {
    id: string;
    sku: string;
    name: string;
    description?: string;
    category: ItemCategory;
    status: ItemStatus;
    unitOfMeasure: UnitOfMeasure;
    currentStock: number;
    reservedStock: number;
    availableStock: number;
    unitCost: number;
    averageCost: number;
    totalValue: number;
    reorderLevel?: number;
    maxLevel?: number;
    minLevel?: number;
    barcode?: string;
    location?: string;
    supplier?: string;
    leadTime: number;
    lastTransactionDate?: Date;
    lastCycleCountDate?: Date;
    createdBy?: string;
    updatedBy?: string;
    createdAt: Date;
    updatedAt: Date;
    stockTransactions: StockTransaction[];
    stockLocations: StockLocation[];
    cycleCountRecords: CycleCount[];
    adjustments: InventoryAdjustment[];
    alerts: InventoryAlert[];
    reorderPoints: ReorderPoint[];
    generateIdentifier(): void;
    isLowStock(): boolean;
    isOutOfStock(): boolean;
    isOverStock(): boolean;
    canFulfillQuantity(quantity: number): boolean;
    getTurnoverRatio(annualUsage: number): number;
    getDaysOfSupply(dailyUsage: number): number;
    getStockStatus(): 'HEALTHY' | 'LOW' | 'OUT_OF_STOCK' | 'OVERSTOCK';
    calculateEOQ(annualDemand: number, orderCost: number, holdingCost: number): number;
    updateAverageCost(newCost: number, quantity: number): void;
    reserve(quantity: number): boolean;
    unreserve(quantity: number): boolean;
    toJSON(): {
        id: string;
        sku: string;
        name: string;
        description: string | undefined;
        category: ItemCategory;
        status: ItemStatus;
        unitOfMeasure: UnitOfMeasure;
        currentStock: number;
        availableStock: number;
        reservedStock: number;
        unitCost: number;
        averageCost: number;
        totalValue: number;
        reorderLevel: number | undefined;
        maxLevel: number | undefined;
        minLevel: number | undefined;
        location: string | undefined;
        supplier: string | undefined;
        leadTime: number;
        stockStatus: "OUT_OF_STOCK" | "OVERSTOCK" | "LOW" | "HEALTHY";
        isLowStock: boolean;
        isOutOfStock: boolean;
        createdAt: Date;
        updatedAt: Date;
    };
}
//# sourceMappingURL=InventoryItem.d.ts.map